clean up and add pipeline
