+++ pipeline/Dockerfile
@@ -1,33 +1,15 @@
-# Start with an official Node image
-FROM node:18
+FROM ghcr.io/osgeo/gdal:alpine-small-latest
 
-# Update the system and install necessary dependencies
-RUN apt-get update && \
-  apt-get install -y parallel jq && \
-  apt-get clean && \
-  rm -rf /var/lib/apt/lists/*
+RUN apk add --no-cache nodejs yarn git
 
-# Create an app directory to hold the application code inside the image
-WORKDIR /usr/src/app
+WORKDIR /app
 
-# Copy your package.json and package-lock.json (if you have one) into the container
-COPY package*.json ./
+COPY package.json ./
+COPY yarn.lock ./
 
-# Install your Node dependencies
-RUN npm install
+RUN yarn install
 
 # Copy your Node scripts into the container
-COPY fetchOsc.js ./
-COPY parser.js ./
-
-# The main script to run the tasks
-COPY process.sh ./
-
-# Install geojson-merge
-RUN npm install -g geojson-merge
-
-# Give execute permissions to the script
-RUN chmod +x process.sh
-
-# The command to run when the container starts
-ENTRYPOINT [ "./process.sh" ]
+COPY src/ ./src/
+COPY cli.js ./
+RUN chmod +x cli.js
