+++ pipeline/package.json
@@ -1,8 +1,7 @@
 {
-  "name": "osmgradient",
+  "name": "osm-gradient-pipeline",
   "version": "1.0.0",
   "description": "Minutely metrics for OSM using FlatGeoBuff",
-  "main": "index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
   },
@@ -11,11 +10,8 @@
   "dependencies": {
     "axios": "^1.5.0",
     "commander": "12.0.0",
-    "osm-adiff-parser": "^1.1.0",
     "real-changesets-parser": "https://github.com/developmentseed/real-changesets-parser.git",
-    "sax": "^1.3.0",
-    "serve": "^14.2.1"
-
+    "sax": "^1.3.0"
   },
   "devDependencies": {
     "geojson-validation": "^1.0.2"
