+++ pipeline/src/config.js
@@ -1,0 +1,5 @@
+const config = {
+    DATA_PATH: process.env.DATA_PATH || '/tmp',
+};
+
+module.exports = config;
